+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.2.5
+ - filesystem security module
+ - unit, action, tool, contact, and service schemas
+ - tasks and message schemas
+
 ## 0.2.4
 
 - add jacsRegistration signature field